Remove standalone regime_features module, consolidate into transform
Moves _MarketRegimeFeatures class into market_regime.py as an internal
implementation detail. Users only interact with MarketRegimeTransform.

Changes:
- Move MarketRegimeFeatures class into market_regime.py
- Rename to _MarketRegimeFeatures (private/internal class)
- Delete standalone torchtrade/envs/offline/regime_features.py
- Update tests to import from transform module
- Simplify test_feature_names (no longer tests get_feature_names method)

Benefits:
- Single source of truth: all regime feature code in one place
- Clearer API: users only see MarketRegimeTransform
- Implementation detail: _MarketRegimeFeatures is internal only
- Simpler imports: from torchtrade.envs.transforms import MarketRegimeTransform

All tests passing:
- ✅ 17 regime feature tests (100% pass)
- ✅ 62 FuturesOneStepEnv tests (100% pass, no regressions)
